requirements.txt
flask==2.3.3
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
pandas>=2.2.0
python-dateutil==2.8.2 